agggg1/ET HACK/agrigenaai/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
groq==0.9.0
streamlit==1.35.0
requests==2.32.0
python-multipart==0.0.9
pydantic==2.7.0
python-dotenv==1.0.1
Pillow==10.3.0
